Fix rank data;
